requirements.txt
python-nmap==0.6.1
openai==0.27.8
rich==13.3.4
requests==2.31.0
scapy==2.6.1
colorama==0.4.6
python-dotenv==1.0.0
openai
python-nmap
requests

